Fixed a bug related to #1 that caused a StackOverflow on filtering
Changed the logic on how custom settings are added to the BurpSuite Settings frame. Tying it to a TreeListener was too volatile. Now it's tied to the "All" filter button. On first calls to "addMenuToSettingsTree" the new menu is injected then anytime the "All" filter button is clicked the custom menus are reinserted.